Change 64 bit ints to 32 bit
We originally used 64 bit ints under the assumption that it wouldn't
really be a problem with modern platforms. However, there have already
been bugs encountered in Emscripten with 64 bit ints, not to mention
that JavaScript doesn't support 64 bit natively (making 64 bit
operations slow).

There was no clear advantage from using 64 bit and using 32 bit avoids
these issues.

+++ compose.c
@@ -21,7 +21,7 @@ static pair ot_compose_skip_skip(ot_comp_skip skip1, size_t offset1,
     size_t skip2_count = (size_t)skip2.count - offset2;
     size_t min_len = min(skip1_count, skip2_count);
 
-    ot_skip(composed, (int64_t)min_len);
+    ot_skip(composed, (uint32_t)min_len);
 
     return (pair) { min_len, min_len };
 }
@@ -51,7 +51,7 @@ static pair ot_compose_skip_delete(ot_comp_skip skip, size_t skip_offset,
     size_t delete_len = (size_t) delete.count - delete_offset;
     size_t min_len = min(skip_len, delete_len);
 
-    ot_delete(composed, (int64_t)min_len);
+    ot_delete(composed, (uint32_t)min_len);
 
     return (pair) { min_len, min_len };
 }
@@ -104,7 +104,7 @@ static pair ot_compose_delete(ot_comp_delete del, size_t del_offset,
                               ot_op* composed) {
 
     size_t del_len = (size_t)del.count - del_offset;
-    ot_delete(composed, (int64_t)del_len);
+    ot_delete(composed, (uint32_t)del_len);
 
     return (pair) { del_len, 0 };
 }
@@ -163,7 +163,7 @@ ot_op* ot_compose(ot_op* op1, ot_op* op2) {
             if (op1_comp == NULL) {
                 assert(!"Both op components should never be NULL.");
             } else if (op1_comp->type == OT_DELETE) {
-                int64_t count = op1_comp->value.delete.count;
+                uint32_t count = op1_comp->value.delete.count;
                 ot_delete(composed, count);
                 op1_next = ot_iter_skip(&op1_iter, (size_t)count);
             } else {
